+++ CHANGES.md
@@ -4,6 +4,7 @@
 * remove OrderUpdate().addTrackingData(trackingData) and replace it with enhanced tracking API
 * add methods to change channel roles
 * integrate syncInfo field into orders
+* add CartState Merged
 
 ## v0.51.0